requirements.txt
Werkzeug
pymime
gevent